Now de-serializes a selected scene on menu item click. Made package accessible to default assembly.
SceneConverter: added to parse a selected scene file when the menu item is clicked. Is currently able to de-seriailze a scene file to package's YamlNodes. Nothing is done with them yet.

YamlDotNet.asmdef: needed to enable autoReferenced to have the package be seeable/usable by the default assembly.

ProjectSetting: updated the scripting language version to 4.5. This might not be needed to get the package to work.