Tried a couple ways of parsing the scene file. Found a mostly good process for doing so.
Tried to deserialize to unity types but received "InvalidCastException: Object must implement IConvertible."

Tried to use derFunk's UnityNodeTypeResolver to convert the yaml to objects. That didn't fix the exception.

One of the authors of YamlDotNet suggested just using the YamlStream to traverse lists and dictionaries that represent the scene. This can work for most parsing of the scene but it doesn't store the guid of a document that is written on the same line as the document start (--- !u!29 &1) so serialized references won't work.

The non-latest attempts to parse the scene file will be removed in a following commit.